BREEZ-1068 Android PA - Add Appointment flow,Fix bug

+++ CarePay/CarePayServiceLibrary/src/main/java/com/carecloud/carepay/service/library/dtos/AvailableLocationDTO.java
@@ -11,7 +11,7 @@ public class AvailableLocationDTO {
 
     @SerializedName("id")
     @Expose
-    private double id;
+    private Integer id;
     @SerializedName("guid")
     @Expose
     private String guid;
@@ -22,7 +22,7 @@ public class AvailableLocationDTO {
     @Expose
     private boolean isVisibleAppointmentScheduler = true;
 
-    public double getId() {
+    public Integer getId() {
         return id;
     }
